% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aggregate.R
\name{add_season}
\alias{add_season}
\title{Add Season Column}
\usage{
add_season(dateVec)
}
\arguments{
\item{rawData}{raw data returned from readNWMdata}
}
\value{
rawData with added season column
}
\description{
Add Season Column
}
\keyword{internal}
