% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aggregate.R
\name{add_waterYear}
\alias{add_waterYear}
\title{Add Water Year Column}
\usage{
add_waterYear(dateVec)
}
\arguments{
\item{dateVec}{raw data returned from readNWMdata}
}
\value{
vector of water years
}
\description{
Add Water Year Column
}
