% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-aggregate.R
\name{aggregate_ys}
\alias{aggregate_ys}
\title{Aggregate by Year-Season}
\usage{
aggregate_ys(rawData, fun = "mean", na.rm = TRUE)
}
\arguments{
\item{rawData}{data extracted with \code{readNWMdata}}

\item{fun}{function to be applied to the flows column default  = 'mean'}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Aggregate by Year-Season
}
\details{
NWM data is extracted as hourly values.

To aggregate hourly data to different time chunks the nwmHistoric package offers a family of aggregate functions.

Each of these begins with the prefix 'aggregate_' and is followed by the date symbol to aggregate to.\tabular{lc}{
   Symbol \tab Aggregate \cr
   y \tab year \cr
   m \tab month \cr
   d \tab day \cr
   j \tab julien day \cr
   s \tab season \cr
   wy \tab water year \cr
}
}
\examples{
\dontrun{
# Get flow record for COMID 101
flows = readNWMdata(comid = 101)

# Aggregate to yearly average (y)
yearly = aggregate_y(flows, fun = 'mean')

# Aggregate to monthly
# minimum and maximum per year (ym)
ym = aggregate_ym(flows, fun = list(min, max))

# Aggregate to seasonal 95th percetile 
# with using custom function
s95 = aggregate_s(flows, fun = function(x){quantile(x,.95)})
}
}
\seealso{
Other aggregate functions: 
\code{\link{aggregate_dowy}()},
\code{\link{aggregate_j}()},
\code{\link{aggregate_m}()},
\code{\link{aggregate_record}()},
\code{\link{aggregate_s}()},
\code{\link{aggregate_wymd}()},
\code{\link{aggregate_wym}()},
\code{\link{aggregate_wys}()},
\code{\link{aggregate_wy}()},
\code{\link{aggregate_yj}()},
\code{\link{aggregate_ymd}()},
\code{\link{aggregate_ym}()},
\code{\link{aggregate_y}()}
}
\concept{aggregate functions}
