% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts-nc.R
\name{check_nco}
\alias{check_nco}
\title{Check to See if NCO is on the system}
\usage{
check_nco()
}
\value{
a boolean condition
}
\description{
NCO is a fantastic open source tool for working with NetCDF files.
It can be downloaded from here: http://nco.sourceforge.net/#Source
}
\examples{
check_nco()
}
