% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts-nc.R
\name{create_nwm_nc}
\alias{create_nwm_nc}
\title{Create a Single NetCDF for list of NWM files}
\usage{
create_nwm_nc(
  fileList = NULL,
  variable = "streamflow",
  dstfile = NULL,
  nco = TRUE,
  pivot = TRUE,
  purge = FALSE
)
}
\arguments{
\item{fileList}{a vector of file paths to merge}

\item{variable}{the channel file variable to extract}

\item{dstfile}{the file path to write to}

\item{nco}{should NCO be used? Default = TRUE}

\item{pivot}{should the concatenated file be pivoted? Default = TRUE}

\item{purge}{should the fileList files be deleted? Default = FALSE}
}
\value{
file path
}
\description{
This function will take a set of NWM files and concatenate them
along the time access. If NCO is installed, this file will be pivoted and 
optimized for time series extraction. While this function will work without NCO
it is highly recommended to install the NCO binaries (http://nco.sourceforge.net/#Source)
}
