% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{crop_flipped_nwm}
\alias{crop_flipped_nwm}
\title{Crop Flipped Raster}
\usage{
crop_flipped_nwm(x, AOI)
}
\arguments{
\item{x}{SpatRast object}

\item{AOI}{a sf polygon}
}
\value{
SpatRast object (x cropped to AOI)
}
\description{
Crop Flipped Raster
}
