% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{download_files}
\alias{download_files}
\title{Download Remote Files}
\usage{
download_files(fileList, outdir = ".")
}
\arguments{
\item{fileList}{fileList object}

\item{outdir}{directory to write files}
}
\value{
data.frame
}
\description{
Download Remote Files
}
