% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasts.R
\name{download_nomads}
\alias{download_nomads}
\title{Download NOMADs files}
\usage{
download_nomads(urls = NULL, dir = NULL)
}
\arguments{
\item{urls}{return from \code{get_nomads_filelist}}

\item{dir}{a directory to write data to}
}
\value{
a list of meta.data information
}
\description{
Download NOMADs files
}
