% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasts.R
\name{error_checking}
\alias{error_checking}
\title{Error Checking for confirugartion/ensemble pairs}
\usage{
error_checking(type, ensemble)
}
\arguments{
\item{type}{a NWM configuration}

\item{ensemble}{an ensemble member number}
}
\value{
a kosher configuration
}
\description{
Error Checking for confirugartion/ensemble pairs
}
