% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts-nc.R
\name{extract_nwm}
\alias{extract_nwm}
\title{Extract NWM flow series}
\usage{
extract_nwm(file, comids, variable = "streamflow")
}
\arguments{
\item{file}{a NetCDF file produced with create_nomads_nc, or create_nomads_ts}

\item{comids}{a vector of NHD comids to extract}

\item{variable}{a channel file variable to extract
(should be same as the one used to create the file)}
}
\value{
a data.frame of comid,dateTime,value
}
\description{
Extract NWM flow series
}
\examples{
\dontrun{
 outfile = create_nwm_nc(type = "short_range", dstfile = tempfile(ext = ".nc"))
 flows = extract_nwm(outfile, comids = c(101,1001,102900))
}
}
