% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasts.R
\name{find_f}
\alias{find_f}
\title{Define 'f'}
\usage{
find_f(type, files, domain)
}
\arguments{
\item{type}{a NWM configuration}

\item{files}{a file list}

\item{domain}{a domain (conus or hawaii)}
}
\value{
a vector of f values.
}
\description{
Define 'f'
}
