% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filelists.R
\name{get_aws_urls}
\alias{get_aws_urls}
\title{Get GCP file list}
\usage{
get_aws_urls(
  version = 2.1,
  output = "CHRTOUT",
  config = NULL,
  ensemble = NULL,
  date = "2010-10-29",
  hour = "00",
  minute = "00",
  num = 3
)
}
\arguments{
\item{version}{NWM model version}

\item{output}{the NWM model output type}

\item{config}{the NWM model configurarion}

\item{ensemble}{the NWM ensemble number}

\item{date}{date of interest}

\item{hour}{hour of interest}

\item{minute}{minute of interest}

\item{num}{number of files to get (forward from provides data-hour-minute)}
}
\value{
data.frame
}
\description{
Get GCP file list
}
