% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filelists.R
\name{get_gcp_urls}
\alias{get_gcp_urls}
\title{Get GCP file list}
\usage{
get_gcp_urls(
  config = "short_range",
  domain = "conus",
  date,
  hour = "00",
  minute = "00",
  num,
  ensemble = NA,
  output = "channel_rt"
)
}
\arguments{
\item{config}{}

\item{domain}{}

\item{date}{}

\item{hour}{}

\item{minute}{}

\item{num}{}

\item{ensemble}{}

\item{output}{}
}
\value{
character vector
}
\description{
Get GCP file list
}
