% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gridded_data}
\alias{get_gridded_data}
\title{Extract Timeseries from fileList}
\usage{
get_gridded_data(fileList, AOI, varname, outfile = NULL, ...)
}
\arguments{
\item{fileList}{a list of gridded NWM outputs}

\item{AOI}{area of interest (sf POLYGON) to subset}

\item{varname}{the name of the variable to extract}
}
\value{
data.frame
}
\description{
Extract Timeseries from fileList
}
