% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filelists.R
\name{get_nomads_filelist}
\alias{get_nomads_filelist}
\title{Get NOMADs File List}
\usage{
get_nomads_filelist(
  config = "short_range",
  domain = "conus",
  date = NULL,
  hour = NULL,
  minute = "00",
  num,
  ensemble = NULL,
  output = "channel_rt",
  version = "prod"
)
}
\arguments{
\item{config}{the NWM model configurarion}

\item{domain}{the NWM model domain}

\item{date}{date of interest}

\item{hour}{hour of interest}

\item{minute}{minute of interest}

\item{num}{number of files to get (forward from provides data-hour-minute)}

\item{ensemble}{the NWM ensemble number}

\item{output}{the NWM model output type}

\item{version}{server version (prod or para)}
}
\value{
data.frame
}
\description{
Get NOMADs File List
}
