% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filelists.R
\name{get_nomads_filelist2}
\alias{get_nomads_filelist2}
\title{Get NOMADs File List}
\usage{
get_nomads_filelist2(
  config = "short_range",
  domain = "conus",
  date = NULL,
  hour = NULL,
  minute = "00",
  num,
  ensemble = NULL,
  output = "channel_rt",
  version = "prod"
)
}
\arguments{
\item{config}{}

\item{domain}{}

\item{date}{}

\item{hour}{}

\item{minute}{}

\item{num}{}

\item{ensemble}{}

\item{output}{}

\item{version}{}
}
\value{
character vector
}
\description{
Get NOMADs File List
}
