% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readNWM.R
\name{get_nwm_meta}
\alias{get_nwm_meta}
\title{NWM metadata}
\usage{
get_nwm_meta(version = NULL)
}
\arguments{
\item{version}{Which version of NWM should be returned? (1.2, 2.0, 2.1)}
}
\value{
data.frame
}
\description{
NWM metadata
}
