% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readNWM.R
\name{get_tds}
\alias{get_tds}
\title{Get TDS path}
\usage{
get_tds(type = "hydroshare")
}
\arguments{
\item{type}{what TDS to use?}
}
\value{
a URL path
}
\description{
Get TDS path
}
\keyword{#}
\keyword{2021}
\keyword{CHANGED}
\keyword{March}
\keyword{http://thredds.hydroshare.org/thredds/dodsC/nwm_retrospective/'}
\keyword{in}
\keyword{internal}
