% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Extract Timeseries from fileList}
\usage{
get_timeseries(
  fileList,
  ids = NULL,
  index_id = "feature_id",
  varname = "streamflow",
  outfile = NULL,
  ...
)
}
\arguments{
\item{fileList}{a list of non-gridded NWM outputs}

\item{ids}{a set of ids to limit the returned data to}

\item{index_id}{the name of the id attributes}

\item{varname}{the name of the variable}
}
\value{
data.frame
}
\description{
Extract Timeseries from fileList
}
