% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{get_timeseries_local}
\alias{get_timeseries_local}
\title{Extract Timeseries from local file list}
\usage{
get_timeseries_local(
  fileList,
  ids = NULL,
  index_id = "feature_id",
  varname = "streamflow",
  outfile = NULL
)
}
\arguments{
\item{fileList}{a list of non-gridded NWM outputs}

\item{ids}{a set of ids to limit the returned data to}

\item{index_id}{the name of the id attributes}

\item{varname}{the name of the variable}

\item{outfile}{file path to save data to (.nc extension)}
}
\value{
data.frame
}
\description{
Extract Timeseries from local file list
}
