% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasts.R
\name{make_url}
\alias{make_url}
\title{make NOMADS url from date and configuration}
\usage{
make_url(date, type)
}
\arguments{
\item{date}{a data in form (YYY-MM-DD)}

\item{type}{a NWM configuration}
}
\value{
a url path
}
\description{
make NOMADS url from date and configuration
}
