% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwmTools.R
\docType{data}
\name{nwm_data}
\alias{nwm_data}
\title{NWM Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 307 rows and 14 columns.
}
\usage{
nwm_data
}
\description{
NWM Data
}
\keyword{datasets}
