% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-readNWM.R
\name{retro_call}
\alias{retro_call}
\title{Retro Call Meta Data}
\usage{
retro_call(comid, meta.obj)
}
\arguments{
\item{comid}{an NHD COMID(s)}

\item{meta.obj}{a metadata object generated with `error.check`}
}
\value{
a data.frame and opend DAP object
}
\description{
build out THREDDs call and meta information
}
\keyword{internal}
