% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts-nc.R
\name{use_nco}
\alias{use_nco}
\title{Use NCO to transform files}
\usage{
use_nco(in_files, variable = "streamflow", dstfile = NULL, pivot = TRUE)
}
\arguments{
\item{in_files}{a vector of file paths}

\item{variable}{the variable to extract}

\item{dstfile}{the location to store the output file}

\item{pivot}{should the 2D variable array be pivoted? Default = TRUE}
}
\value{
a file path
}
\description{
Workflow for transforming set of files to a single, pivoted, 
NetCDF file using NCO (see: check_nco).
}
