% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts-nc.R
\name{use_r}
\alias{use_r}
\title{Use R to transform files}
\usage{
use_r(in_files, variable = "streamflow", dstfile = NULL)
}
\arguments{
\item{in_files}{a vector of file paths}

\item{variable}{the variable to extract}

\item{dstfile}{the location to store the output file}
}
\value{
a file path
}
\description{
Workflow for transforming set of files to a single, 
NetCDF file using RNetCDF.
}
