% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gridded_data}
\alias{get_gridded_data}
\title{Extract Gridded Data from fileList}
\usage{
get_gridded_data(fileList, AOI, varname, outfile = NULL)
}
\arguments{
\item{fileList}{a list of gridded NWM outputs}

\item{AOI}{area of interest (sf POLYGON) to subset}

\item{varname}{the name of the variable to extract}

\item{outfile}{filepath to save data (with .nc extension)}
}
\value{
data.frame
}
\description{
Extract Gridded Data from fileList
}
