% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasts.R
\name{get_nomads_filelist}
\alias{get_nomads_filelist}
\title{Build file metadata for most current NWM configuration}
\usage{
get_nomads_filelist(type = NULL, ensemble = NULL, num = 6)
}
\arguments{
\item{type}{a NWM configuration}

\item{ensemble}{an ensemble member number}

\item{num}{a number of files to download, default = all}
}
\value{
a list of meta.data information
}
\description{
Build file metadata for most current NWM configuration
}
\examples{
\dontrun{
 urls     <-  get_nomads_filelist(type = "medium_range", num = 20,  ensemble = 4)
 in_files <-  download_nomads(urls, dir = tempdir())
}
}
